ucp-store-mocker/pyproject.toml
[project]
name = "ucp-store-mocker"
version = "0.1.0"
description = "CLI tool for generating UCP-compliant mock stores for developer enablement"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
keywords = ["ucp", "commerce", "mock", "testing", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.0",
    "faker>=20.0",
    "rich>=13.0",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
ucp-store-mocker = "ucp_store_mocker.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ucp_store_mocker"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
